Refine domain and data package organization
