+++ app/src/main/java/com/gabrielthecode/kontakt/data/local/dao/RemoteKeyDao.kt
@@ -1,0 +1,28 @@
+package com.gabrielthecode.kontakt.data.local.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import com.gabrielthecode.kontakt.data.local.entities.RemoteKeyEntity
+
+@Dao
+interface RemoteKeyDao {
+	@Insert(onConflict = OnConflictStrategy.REPLACE)
+	suspend fun insertKeys(keys: List<RemoteKeyEntity>)
+
+	@Insert(onConflict = OnConflictStrategy.REPLACE)
+	suspend fun insertRemoteKey(key: RemoteKeyEntity)
+
+	@Query("SELECT * FROM remote_key where id=:id")
+	suspend fun getRemoteKeyById(id: String): RemoteKeyEntity?
+
+	@Query("SELECT * FROM remote_key")
+	suspend fun getAllRemoteKeys(): List<RemoteKeyEntity>
+
+	@Query("DELETE FROM remote_key")
+	suspend fun clearRemoteKeys()
+
+	@Query("Select createdAt From remote_key Order By createdAt DESC LIMIT 1")
+	suspend fun getCreationTime(): Long?
+}
