+++ app/src/main/java/com/gabrielthecode/kontakt/data/local/entities/RemoteKeyEntity.kt
@@ -1,0 +1,14 @@
+package com.gabrielthecode.kontakt.data.local.entities
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "remote_key")
+data class RemoteKeyEntity(
+	@PrimaryKey(autoGenerate = false)
+	val id: String,
+	val prevKey: Int?,
+	val currentPage: Int,
+	val nextKey: Int?,
+	val createdAt: Long = System.currentTimeMillis()
+)
