--- app/src/main/java/com/gabrielthecode/kontakt/data/network/RandomUserApiRemoteService.kt
@@ -1,9 +1,0 @@
-package com.gabrielthecode.kontakt.data.network
-
-import com.gabrielthecode.kontakt.data.network.model.UserResponse
-
-interface RandomUserApiRemoteService {
-	suspend fun getUserContacts(
-		page: Int
-	): UserResponse
-}
