--- app/src/main/java/com/gabrielthecode/kontakt/data/network/mapper/EntityMapper.kt
@@ -1,6 +1,0 @@
-package com.gabrielthecode.kontakt.data.network.mapper
-
-interface EntityMapper<Entity, DomainModel> {
-	fun mapToDomain(entity: Entity): DomainModel
-	fun mapToEntity(domainModel: DomainModel): Entity
-}
