+++ app/src/main/java/com/gabrielthecode/kontakt/data/remote/RandomUserApiRemoteService.kt
@@ -1,0 +1,9 @@
+package com.gabrielthecode.kontakt.data.remote
+
+import com.gabrielthecode.kontakt.data.remote.model.UserResponse
+
+interface RandomUserApiRemoteService {
+	suspend fun getUserContacts(
+		page: Int
+	): UserResponse
+}
