+++ app/src/main/java/com/gabrielthecode/kontakt/data/remote/datasource/RemoteDataSource.kt
@@ -1,0 +1,9 @@
+package com.gabrielthecode.kontakt.data.remote.datasource
+
+import com.gabrielthecode.kontakt.domain.model.UserDomainModel
+
+interface RemoteDataSource {
+	suspend fun getUserContacts(
+		page: Int
+	): UserDomainModel
+}
