+++ app/src/main/java/com/gabrielthecode/kontakt/data/remote/mapper/EntityMapper.kt
@@ -1,0 +1,6 @@
+package com.gabrielthecode.kontakt.data.remote.mapper
+
+interface EntityMapper<Entity, DomainModel> {
+	fun mapToDomain(entity: Entity): DomainModel
+	fun mapToEntity(domainModel: DomainModel): Entity
+}
