--- app/src/main/java/com/gabrielthecode/kontakt/domain/remote/RemoteDataSource.kt
@@ -1,9 +1,0 @@
-package com.gabrielthecode.kontakt.domain.remote
-
-import com.gabrielthecode.kontakt.domain.model.UserDomainModel
-
-interface RemoteDataSource {
-	suspend fun getUserContacts(
-		page: Int
-	): UserDomainModel
-}
