+++ app/src/main/java/com/gabrielthecode/kontakt/domain/usecases/GetUserContacts.kt
@@ -1,8 +1,8 @@
 package com.gabrielthecode.kontakt.domain.usecases
 
 import androidx.paging.PagingData
-import com.gabrielthecode.kontakt.domain.repositories.UserContactsRepository
-import com.gabrielthecode.kontakt.data.database.entities.UserContactEntity
+import com.gabrielthecode.kontakt.data.repositories.UserContactsRepository
+import com.gabrielthecode.kontakt.data.local.entities.UserContactEntity
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
