+++ app/src/test/java/com/gabrielthecode/kontakt/UserContactRemoteMediatorTest.kt
@@ -8,10 +8,10 @@ import androidx.paging.RemoteMediator
 import com.gabrielthecode.kontakt.domain.model.Info
 import com.gabrielthecode.kontakt.domain.model.UserContactDomainModel
 import com.gabrielthecode.kontakt.domain.model.UserDomainModel
-import com.gabrielthecode.kontakt.domain.local.LocalDataSourceImpl
-import com.gabrielthecode.kontakt.domain.remote.RemoteDataSourceImpl
-import com.gabrielthecode.kontakt.data.database.mapper.UserEntityToDomainMapper
-import com.gabrielthecode.kontakt.data.database.entities.UserContactEntity
+import com.gabrielthecode.kontakt.data.local.datasource.LocalDataSourceImpl
+import com.gabrielthecode.kontakt.data.remote.datasource.RemoteDataSourceImpl
+import com.gabrielthecode.kontakt.data.local.mapper.UserEntityToDomainMapper
+import com.gabrielthecode.kontakt.data.local.entities.UserContactEntity
 import com.gabrielthecode.kontakt.data.mediator.UserContactRemoteMediator
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
