+++ app/src/test/java/com/gabrielthecode/kontakt/UserContactsViewModelTest.kt
@@ -2,7 +2,7 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.paging.PagingData
 import androidx.paging.map
 import com.gabrielthecode.kontakt.domain.usecases.GetUserContacts
-import com.gabrielthecode.kontakt.data.database.entities.UserContactEntity
+import com.gabrielthecode.kontakt.data.local.entities.UserContactEntity
 import com.gabrielthecode.kontakt.presentation.contact.UserContactEvent
 import com.gabrielthecode.kontakt.presentation.contact.UserContactsViewModel
 import com.gabrielthecode.kontakt.presentation.contact.mapper.UserContactEntityToUIModelMapper
